Introduce the new API: psh_open_session_with_name(char *name)
BZ: 172529

Currently psh_open_session(psh_sensor_t sensor_type) is used to open a session with sensorhubd.
psh_sensor_t is enumeration.The shortcoming of it is, when a new sensor type need to be supported,
the source code of libsensorhub have to be modified and re-compiled.

So we introduce a new API, psh_open_session_with_name(char *name), name is a UID string to identify
the sensor type, that is defined inside PSH FW. When a new sensor type need to be supported,
not a single line of libsensorhub code would be changed with this approach.

Change-Id: Ia0c6f902a7f71a35f857ca6feabe31d138624f1c
Signed-off-by: Forrest Zhao <forrest.zhao@intel.com>